Finished Label and font part of ResourceManager
Threw a ResourceManager into the Game and fixed things a bit up in the
occasion.
Moved things around and simpli/prettified Game.
Fixed up the Label, like it's Constructor.
Hooked the label font loading up to the ResourceManager.
Made and sorta tested the font functionality of the ResourceManager.

Declaring Label and the font part of the Resource Manager done. Now the
only reason to fiddle with it is bugs and additional accessors. Unless i
decide to reimplement it all again, which i won't. It works fine right
now already, I think.