Slack from school, Hardcore programming.
Redid the entire Input system:
Made the Input into an InputManager.
Based it in the event system.
Mane the Menu folder into MenuControls.
Removed the static MainMenu function and made a Menu deriving MainMenu
class.
Made a DoTick in the Scene class that calls a virtual Tick class.
And a ton more stuff.