Christmas work
Extended the resourcemanagement to keep track on how long a resource is
kept in the system. Persistance and such.
Made the resourcemanager clear resources depending on the passed depth
message.
Made scene change happen in the start of the loop and not
instantaniouly.
Made a pure virtual Load function for the Scene for the scenes to load
in instead of in the constructor.
Fixed the background.
Prettified things here and there.
Added an OptionsMenu.
Made an initial extendable game folder architecture.